Generating models with large number of blocks and S-functions
