+++ slsf/testrun.m
